fix duplicate error message
which came from using the id field as the doc id for the corpus function. The id field for the tweets somehow lost data--after a certain point it spit out a bunch of zeros as seen in the CSV file. In the CSV, and I assume in the R data as well, the id_str field retains those digits and they are not switched to zeros. Relying on that field sovles the duplicate problem.